add purchase_price column to pivot table

+++ app/Http/Controllers/CoinsController.php
@@ -8,7 +8,7 @@ use Illuminate\Support\Facades\DB;
 class CoinsController extends Controller
 {
     public function index(){
-        $thisUserId = auth('sanctum')->id();
+        $thisUserId = (int) auth('sanctum')->id();
         return DB::table('coins')
             ->leftJoin('user_coin', function($join) use ($thisUserId) {
                 $join->on('user_coin.coin_id', '=', 'coins.id')
